+++ LeaveManagement.Application/Features/LeaveTypes/Handlers/Commands/DeleteLeaveTypeCommandHandler.cs
@@ -1,5 +1,7 @@
-﻿using HRLeaveManagement.Application.Features.LeaveTypes.Requests.Commands;
+﻿using HRLeaveManagement.Application.Exceptions;
+using HRLeaveManagement.Application.Features.LeaveTypes.Requests.Commands;
 using HRLeaveManagement.Application.Persistence.Contracts;
+using HRLeaveManagement.Domain;
 using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
@@ -17,6 +19,10 @@ namespace HRLeaveManagement.Application.Features.LeaveTypes.Handlers.Commands
         public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
         {
             var leaveType = await _leaveTypeRepository.Get(request.Id);
+
+            if (leaveType == null)
+                throw new NotFoundException(nameof(LeaveType), request.Id);
+
             await _leaveTypeRepository.Delete(leaveType);
             return Unit.Value;
         }
